feat(improvement): better data selection for triplet margin loss
Follow more the logic of triplet margin loss by finding a similar
example(positive) and a very different counter example(negative).
However it is still not working as expected and the performance are
really poor as it is.
A good plan would be to cluster the samples to make the search faster
without losing to much quality in the sample selection.
And make a more thorought triplet margin loss on not only counter
example but actual example that are not recognized properly.
That might require -again- a rearchitecture of how I handle the confused
sample class